Point out working directory in step 4 of tutorial
Since step 3 is about files inside the "hello" directory, it is not
immediately obvious that in step 4 "flatpak build-export" needs to be
executed in the parent directory of "hello/". This might lead to some
frustrating debugging or the impression that the tutorial/documentation
is outdated or flatpak is broken. The additional instruction hopefully
prevents this.