getting: Add Arch opt-deps as needed for flatpak-builder
Arch doesn't install opt-depts by default, so flatpak-builder is broken on those system for a simple sudo pacman -S flatpak. Ideally we would need this, but it's Arch decision to not install opt-deps y default.

So document that they need to install it so developers of Flatpak applications doesn't have the problem of eu-strip not working or patch not working.