tweak link names and addresses
It's better to expose people to the flathub home page rather than
directing them straight to the apps - that way they are
encouraged to follow the setup guide before attempting to install
anything.

Use "install" rather than "browse" when linking to flathub, since
we don't have browsing yet.

Consistently refer to the docs as the "Developer guide".

Consistent capitalisation.