Update elementary OS instructions for 5.1+
As of elementary OS 5.1 up-to-date Flatpak is included in the repo and preinstalled, AppCenter supports Flatpak ootb, and the preinstalled Sideload app handles .flatpakref files. Since 5.0 users are upgraded to 5.1 and older versions are EOL, there's no need for older instructions.

Installing an app from a .flatpakref with Sideload also adds its remote, so there's no need to add it separately.